requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
faiss-cpu
